fixes to resolve bug#1
Enables adjustments to Elasticsearch default settings for indexing and searches
by subclassing the ElasticsearchSearchBackend (see search_backend.py). Some
changes also required to settings.py and config.py to enable the new backend.

A few other configuration changes: we can no longer use whoosh or sqlite for
development if we want to be able to test the Elasticsearch adjustments. I am
using a VMware instance that mimics the AWS configuration for Elasticsearch,
and I installed a copy of PostgreSQL on the host (my MacBook Pro) as a
development database.

Checking this in before getting on a plane. Ya know, just in case...