work in progress on interface
The current state of things points the interface at the new search API and
shows all experiments *and* samples matching the search terms. Next up are
detail views for Experiment and Sample and a way to select and add them to
a "shopping cart" (a one-off "experiment") for further analysis.