modify Sample to use autoincrement id, move cel_file
`Sample.sample` is now `Sample.name`, and no longer a primary key. `Sample.id`
is now generated by Django as the default (autoincrement id) and is what will
appear in the API references.

`SampleAnnotation.cel_file` is now `Sample.ml_data_source`. Moving it over to
`Sample` to reflect that this field comes from ArrayExpress and to make it
easier to use when linking with the "activity" spreadsheet.

No other changes in SampleAnnotation yet, so API still works as before except
that `Sample` will also now return the `ml_data_source` when retrieved.