enforce unique constraint on ml_data_source
All changes related to ensuring that ml_data_source remains unique across all
Samples created in the database. The tricky part here is that we also allow
this field to be NULL for any Sample that does not have any publicly-available
reference data. Django handles this, but we need to specify the right flags
everywhere to explicitly state that ml_data_source may be null. Haystack also
needed the memo about that or it got cranky about indexing, hence the inclusion
of search_indexes.py here.

Note that I have modified the file naming in the data/ directory to facilitate
tracking changes in the annotation data across versions.