initial changes to SampleAnnotation model
SampleAnnotation is now more flexible and can take arbitrary name, value
annotations for any Sample with the constraint that, within a given sample
instance, each sampleannotation.name must be unique. (Only one annotation of
a given AnnotationType allowed.)

Data are tested to load as expected, but api.py and search index should be
broken for now.