Fix glitch in test_get_annotations_get_param and elsewhere.
This fixes the sporadic problem we encountered when fixtureless happened to
generate test data that didn't get URL encoded properly, causing an unwarranted
testing failure.

TastyPie's api_client.get() uses a separate input parameter
(data=) to handle URL parameters. Although it works sometimes for us to
manually assemble a complete URI in the testing code, we are also then
responsible for properly URL-encoding the parameters. This set of changes now
calls api_client.get() with a data= object with those URL parameters to let
TastyPie handle this step for us.