Introduce Browserify & Babel to package ttest for our use
The library we are using for ttests was written as a node module using ES2015
syntax, so we have to do some extra work to prepare it for use in the browser.
Browserify packages up the node dependencies into one file for use, while
Babel transpiles from ES2015 syntax to ES5, which makes the whole bundle
compatible with the rest of our toolchain (both uglify and PhantomJS had
issues with the newer ES2015 constructs) as well as browsers that don't yet
fully support ES2015.