Get codeclimate to skip over vega spec formatting (*see log).
The vega spec format is JSON, but we need it embedded in our Angular app. This
causes ESLint to parse the spec as ES5, which it sort-of fits, but doesn't quite
meet the style guide, generating many useless errors about string delimiters.

To address this, I've separated the spec into a thinly-wrapped Angular module
with a name ending in *-vgspec.js and tried to get that file glob excluded from
what codeclimate checks.

Another solution for this would be to just reformat the spec to fit our style
guide. I tried that, but then the vega editor complained about the formatting.
I think it's better to be able to use the vega editor when necessary without
any conversion.

If this works, we can use the same method in other places where we need to
embed vega specs.