Flip the sign of the "diff" operation.
It makes more sense to consider "group a" as a base group (usually defined
first) and "group b" as a comparison (usually defined second), so this order
of operations puts the nodes that "increase" activity relative to base toward
the right and those that "decrease" toward the left.

Many thanks to Jie for the helpful insight here!