Add button to launch network with genes from selected nodes.
This set of changes adds a "View in Gene Network" button to the volcano plot
view when at least one node is selected. Clicking the button will query the
`participation` API for a list of genes participating in the selected nodes
and then launch the network view with that list of genes as well as the
MLModel and sample groups (base vs. comparison) used to generate the volcano
plot.

MLModel is currently defaulted to 1. Need to retro-fit model selection into the
heatmap next.