[FEAT] make POST operate as GET for ExperessionValueResource
This set of changes allow us to invoke calls to the list endpoint for
ExpressionValueResource as GET and POST interchangeably. This works around the
inherent limitations of GET, which limit us to parameters that will fit in the
length limitations of a URI by using the body of a POST message to send those
parameters.

The implementation works by overriding the `post_list` method, which is what
Tastypie usually invokes to update the data backing a list endpoint, and re-
directing that POST request as if it had been a typical Tastypie GET.