[FIX] add tests for Sample indexing, update index
In the course of writing up a test for this, I discovered that our original
implementation of the Sample index was done incorrectly. This arose from a
conceptual error (on my part) in how the Haystack indexes work. The Experiment
accession numbers at the heart of this fix were never getting into the "full
text" index because they were not in the sample_text.txt template! The fields
defined in search_indexes.py (which I thought were being searched) are only
there for filtering and sorting purposes -- and we aren't currently using them
at all.

This fix repairs the indexing process for the SampleIndex.experiments field
in case we decide to use it. Also included is a test which causes a Haystack
index to be built for Samples, which will catch an error like this earlier in
the future. In addition, the full text index for Samples is now updated to
include a list of corresponding Experiment accession numbers.