[FIX] make filter work properly everywhere
There was a bug in the display of sample annotations that was hidden by the
fact that Angular would display an undefined value as a blank table cell.
Adding our 'N/A' filter caused this bug to manifest as uninterpreted template
text in cells that had missing data due to undefined annotations.

This set of changes fixes that problem by creating a "safe" accessor for
annotations that will return an empty string when an annotation type is
not found.