[FIX] improve error reporting for Volcano Plot
Volcano plot cannot be drawn if a sample is missing activity, similar to the
heatmap (issue #253). This change adds rudimentary error reporting to help
explain to a user what is going on when this happens.

Note: this overall process could be better. Ideally, a user would see upfront
when selecting samples that activity data were unavailable and actions that
add those samples to groups would be disabled with an explanation.