[FEAT] clarify when filters are applied, add tip to download menu
This adds a line of text describing how the filters affect the visible network
data in order to clarify what is shown. Also adds a tool tip to the download
menu to emphasize that the downloaded data will match what is shown on the page.

Many thanks to @dongbohu for the review feedback. I think this greatly improves
the page without making it more complicated.